Improve type safety (esp TileID/MarkID/PieceID, board serial numbers, and CPtrArray)
add GShr/CyberBoard.h with various helper definitions

Miscellaneous type safety work

  1) Use size_t for memory allocation size.

  2) CGmBoxHint

  3) WM_OVERRIDE_SELECTED_ITEM parameter

  4) Separate CEventMessageRcd overloaded members into a union to
     allow more precise types

  5) Improve type safety of CProjListBox group code use

     Make CProjListBox a template that uses the specific enum type
     instead of generic int in its public interface.  (The internal
     implementation is unchanged; this just clarifies and enforces
     the intended use of the public interface.)

  6) Mark some virtual functions with override

     This will make it easier to change the signature of these
     functions by pointing out when an override and the base don't
     match.

use XxxxIDTable<> to prepare for using XxxxID<>

  Factor some repetitive code for tables indexed by ID values into
  a new XxxxIDTable<>.

improve type safety of TileID/MarkID/PieceID

  Replace typedef unsigned short TileID/MarkID/PieceID which the
  compiler treats as all being the same type with template XxxxID<>
  so the compiler treats them as different types.

improve type safety of board serial numbers

  Change simple int/WORD board serial numbers implementation to
  typedef XxxxID<'B'> BoardID

replace CPtrArray with std::vector<> for type safety (tiles)

replace CPtrArray with std::vector<> for type safety (marks)

replace CPtrArray with std::vector<> for type safety (pieces)

replace CPtrArray with std::vector<> for type safety (boards)

replace CPtrArray with std::vector<> for type safety (move history)

use size_t for CProjListBox m_nMarkSourceCode

  It matches more of the callers' nSourceCode types

replace CPtrArray with std::vector<> for WinState.*

replace CPtrArray with std::vector<> for LBoxGfx2 family

replace CPtrArray with std::vector<> for trays

miscellaneous replacement of CPtrArray with std::vector<>

improve type safety of CGrafixListBox

  Split the ID-specific parts of CGrafixListBox out into a template
  class that can enforce correct ID type use.  The template class
  then (possibly indirectly) derives from CGrafixListBox to provide
  the non-ID-specific CGrafixListBox code.

fixup! be more explicit about nature of ObjectID

+++ GM/DlgPnew.h
@@ -58,7 +58,7 @@ public:
 
     // ..Must be set by caller..
     CGamDoc*    m_pDoc;             // Pointer to document
-    int         m_nPSet;            // Number of piece set
+    size_t      m_nPSet;            // Number of piece set
 
 // Implementation
 protected:
